+++ src/test/java/com/github/edipermadi/security/blobfish/pool/BlobPoolBuilderTest.java
@@ -176,8 +176,8 @@ public final class BlobPoolBuilderTest extends AbstractTest {
             final Map<UUID, Blob.SimplifiedMetadata> entries = blobPool.listAvailableBlobs(page, 10);
             for (final Map.Entry<UUID, Blob.SimplifiedMetadata> entry : entries.entrySet()) {
                 final UUID blobId = entry.getKey();
-                final Set<String> tags = blobPool.getBlobTags(blobId);
-                Assert.assertTrue(tags.contains(tagVal));
+                final Map<UUID, String> tags = blobPool.getBlobTags(blobId);
+                Assert.assertTrue(tags.containsValue(tagVal));
             }
             empty = entries.isEmpty();
         }
@@ -204,8 +204,8 @@ public final class BlobPoolBuilderTest extends AbstractTest {
             final Map<UUID, Blob.SimplifiedMetadata> entries = blobPool.listAvailableBlobs(page, 10);
             for (final Map.Entry<UUID, Blob.SimplifiedMetadata> entry : entries.entrySet()) {
                 final UUID blobId = entry.getKey();
-                final Set<String> tags = blobPool.getBlobTags(blobId);
-                Assert.assertFalse(tags.contains(tagVal));
+                final Map<UUID,String> tags = blobPool.getBlobTags(blobId);
+                Assert.assertFalse(tags.containsValue(tagVal));
             }
             empty = entries.isEmpty();
         }
@@ -237,8 +237,8 @@ public final class BlobPoolBuilderTest extends AbstractTest {
             final Map<UUID, Blob.SimplifiedMetadata> entries = blobPool.listAvailableBlobs(page, 10);
             for (final Map.Entry<UUID, Blob.SimplifiedMetadata> entry : entries.entrySet()) {
                 final UUID blobId = entry.getKey();
-                final Set<String> tags = blobPool.getBlobTags(blobId);
-                Assert.assertTrue(tags.contains(tagVal));
+                final Map<UUID,String> tags = blobPool.getBlobTags(blobId);
+                Assert.assertTrue(tags.containsValue(tagVal));
             }
             empty = entries.isEmpty();
         }
@@ -252,8 +252,8 @@ public final class BlobPoolBuilderTest extends AbstractTest {
             final Map<UUID, Blob.SimplifiedMetadata> entries = blobPool.listAvailableBlobs(page, 10);
             for (final Map.Entry<UUID, Blob.SimplifiedMetadata> entry : entries.entrySet()) {
                 final UUID blobId = entry.getKey();
-                final Set<String> tags = blobPool.getBlobTags(blobId);
-                Assert.assertFalse(tags.contains(tagVal));
+                final Map<UUID,String> tags = blobPool.getBlobTags(blobId);
+                Assert.assertFalse(tags.containsValue(tagVal));
             }
             empty = entries.isEmpty();
         }
@@ -269,14 +269,19 @@ public final class BlobPoolBuilderTest extends AbstractTest {
             final Map<UUID, Blob.SimplifiedMetadata> blobs = blobPool.listAvailableBlobs(page, 10);
             for (final Map.Entry<UUID, Blob.SimplifiedMetadata> entry : blobs.entrySet()) {
                 final Blob.SimplifiedMetadata metadata = entry.getValue();
-                final Set<String> tags = blobPool.getBlobTags(entry.getKey());
+                final Map<UUID,String> tags = blobPool.getBlobTags(entry.getKey());
                 final byte[] payload = blobPool.getBlobPayload(entry.getKey());
+
+                final Set<String> tagValues = new HashSet<>();
+                for(final Map.Entry<UUID, String> tag: tags.entrySet()){
+                    tagValues.add(tag.getValue());
+                }
 
                 log("found entry");
                 log("  uuid      : %s", entry.getKey());
                 log("  mime-type : %s", metadata.getMimeType());
                 log("  path      : %s", metadata.getPath());
-                log("  tags      : %s", Joiner.on(", ").join(tags));
+                log("  tags      : %s", Joiner.on(", ").join(tagValues));
 
                 final File file = new File(metadata.getPath());
                 final File dir = new File("target");